+++ CHANGELOG.md
@@ -30,6 +30,7 @@ also non string parameters (e.g. `Enum.join([1, 2], ",")`
 - Support for `binary:split/3` and `string:find/2,3`
 - Support for large tuples (more than 255 elements) in external terms.
 - Support for `io:put_chars/2`
+- Support for `lists:nthtail/2`
 
 ### Changed